protect queries, now they can only be executed by logged users, add logic to edit questions and add/delete new options

+++ app/graphql/mutations/delete_option.rb
@@ -4,6 +4,9 @@ class Mutations::DeleteOption < Mutations::BaseMutations
   field :errors, [String], null: false
 
   def resolve(id:)
+    if context[:current_user].blank?
+      raise GraphQL::ExecutionError, 'Authentication required'
+    end
     option = Option.find(id)
     if option.destroy
       {
